# benchmarks/locobench_agent/tasks/csharp_blockchain_defi_expert_070_architectural_understanding_expert_01/environment/project/docs/api/rest/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UtilityChain Core Suite – REST API",
    "description": "OpenAPI definition for the UtilityChain Core Suite monolith.  \nThis gateway exposes wallet, staking, governance, NFT and smart–contract features through a unified REST interface.",
    "version": "1.0.0",
    "termsOfService": "https://utilitychain.local/terms",
    "contact": {
      "name": "UtilityChain Engineering",
      "url": "https://utilitychain.local",
      "email": "support@utilitychain.local"
    },
    "license": {
      "name": "Business-Source License 1.1",
      "url": "https://utilitychain.local/license"
    }
  },
  "servers": [
    {
      "url": "https://api.utilitychain.local",
      "description": "Production"
    },
    {
      "url": "https://sandbox.utilitychain.local",
      "description": "Sandbox"
    }
  ],
...
  "paths": {
...
    "/api/v1/transactions": {
      "post": {
        "tags": [ "Wallet" ],
        "summary": "Submit a transfer transaction",
        "security": [ { "bearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransactionRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Transaction accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransactionAccepted" }
              }
            }
          },
          "400": { "description": "Invalid transaction payload" }
        }
      }
    },
    "/api/v1/transactions/{hash}": {
      "get": {
        "tags": [ "Wallet" ],
        "summary": "Get transaction status",
        "parameters": [
          { "$ref": "#/components/parameters/TxHashParam" }
        ],
        "responses": {
          "200": {
            "description": "Transaction record",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Transaction" }
              }
            }
          },
          "404": { "description": "Transaction not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "TxHashParam": {
        "name": "hash",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "pattern": "^[A-Fa-f0-9]{64}$" },
        "description": "Transaction hash"
      },
...
    },
    "schemas": {
...
      "TransactionRequest": {
        "type": "object",
        "properties": {
          "from": { "type": "string" },
          "to": { "type": "string" },
          "amount": { "type": "string", "example": "50.0000" },
          "fee": { "type": "string", "example": "0.0100" },
          "memo": { "type": "string", "maxLength": 256 }
        },
        "required": [ "from", "to", "amount" ]
      },
      "TransactionAccepted": {
        "type": "object",
        "properties": {
          "hash": { "type": "string" },
          "acceptedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "hash": { "type": "string" },
          "blockHeight": { "type": "integer" },
          "timestamp": { "type": "string", "format": "date-time" },
          "from": { "type": "string" },
          "to": { "type": "string" },
          "amount": { "type": "string" },
          "fee": { "type": "string" },
          "status": { "type": "string", "enum": [ "Pending", "Confirmed", "Failed" ] }
        },
        "required": [ "hash", "from", "to", "amount", "status" ]
      },
...
    }
  }
}